+++ src/apps/spotify/app/poll_events.rs
@@ -9,7 +9,7 @@ pub async fn poll_events<F, Fut>(
     mut in_receiver: Receiver<In>,
     play_or_pause: F,
 ) where
-    F: Fn(Arc<State>, u16) -> Fut + Copy,
+    F: Fn(Arc<State>, usize) -> Fut + Copy,
     Fut: Future<Output = ()>,
 {
     while let Some(event) = in_receiver.recv().await {
@@ -23,7 +23,7 @@ pub async fn poll_events<F, Fut>(
 }
 
 async fn handle_event<F, Fut>(state: Arc<State>, play_or_pause: F, event: In) where
-    F: Fn(Arc<State>, u16) -> Fut,
+    F: Fn(Arc<State>, usize) -> Fut,
     Fut: Future<Output = ()>,
 {
     match event {
@@ -67,7 +67,7 @@ mod test {
             out_sender,
         );
 
-        async fn play_or_pause(state: Arc<State>, _: u16) {
+        async fn play_or_pause(state: Arc<State>, _: usize) {
             state.sender.send(Out::Server(ServerCommand::SpotifyPlay {
                 track_id: "spotify:track:68d6ZfyMUYURol2y15Ta2Y".to_string(),
                 access_token: "access_token".to_string(),
@@ -105,7 +105,7 @@ mod test {
             out_sender,
         );
 
-        async fn play_or_pause(state: Arc<State>, _: u16) {
+        async fn play_or_pause(state: Arc<State>, _: usize) {
             state.sender.send(Out::Server(ServerCommand::SpotifyPlay {
                 track_id: "spotify:track:68d6ZfyMUYURol2y15Ta2Y".to_string(),
                 access_token: "access_token".to_string(),
@@ -135,7 +135,7 @@ mod test {
         let (out_sender, mut out_receiver) = tokio::sync::mpsc::channel::<Out>(32);
         let state = get_state_with_last_action_and_sender(Instant::now(), out_sender);
 
-        async fn play_or_pause(state: Arc<State>, _: u16) {
+        async fn play_or_pause(state: Arc<State>, _: usize) {
             state.sender.send(Out::Server(ServerCommand::SpotifyPlay {
                 track_id: "spotify:track:68d6ZfyMUYURol2y15Ta2Y".to_string(),
                 access_token: "access_token".to_string(),
@@ -164,7 +164,7 @@ mod test {
         let (out_sender, mut out_receiver) = tokio::sync::mpsc::channel::<Out>(32);
         let state = get_state_with_last_action_and_sender(Instant::now() - Duration::from_millis(5_000), out_sender);
 
-        async fn play_or_pause(state: Arc<State>, index: u16) {
+        async fn play_or_pause(state: Arc<State>, index: usize) {
             state.sender.send(Out::Server(ServerCommand::SpotifyPlay {
                 track_id: format!("spotify:track:{}", index),
                 access_token: "access_token".to_string(),
